backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
anthropic
openai
groq
chromadb
pydantic
slowapi
bcrypt
PyJWT
